Replace win32 OS layer with mswin crate
