+++ Cargo.toml
@@ -59,7 +59,7 @@ rust_netbeans = { path = "rust_netbeans" }
 default = []
 
 [target.'cfg(windows)'.dependencies]
-rust_os_win32 = { path = "rust_os_win32" }
+rust_os_mswin = { path = "rust_os_mswin" }
 rust_winclip = { path = "rust_winclip" }
 rust_xpm_w32 = { path = "rust_xpm_w32" }
