+++ rust_os_mswin/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+windows = { version = "0.52", features = ["Win32_System_SystemInformation", "Win32_System_Console"] }
 
 [lib]
 name = "rust_os_mswin"
