--- rust_os_win32/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "rust_os_win32"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-windows = { version = "0.52", features = ["Win32_System_SystemInformation", "Win32_System_Console"] }
-
-[lib]
-name = "rust_os_win32"
-crate-type = ["staticlib", "rlib"]
